+++ packages/crayon/src/plugin/__tests__/resolveGenerator.test.ts
@@ -1,8 +1,8 @@
-import resolveGenerator from './resolveGenerator'
-import * as generators from './generators'
-import { Generator } from '../types'
+import resolveGenerator from '../resolveGenerator'
+import * as generators from '../generators'
+import { Generator } from '../../types'
 
-jest.mock('./tailwindUtils')
+jest.mock('../tailwindUtils')
 
 const generatorsList: Array<[keyof typeof generators, Generator]> = [
   ['backgroundColor', generators.backgroundColor],
